apps/ai/pyproject.toml
[project]
name = "manthanein-ai"
version = "1.0.0"
description = "AI service for Manthanein - document parsing, embeddings, and intelligent features"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    # AI/ML
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "transformers>=4.37.0",
    "sentence-transformers>=2.3.0",
    "torch>=2.1.0",
    # Vector DB
    "qdrant-client>=1.7.0",
    # Document parsing
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.23",
    "markdown>=3.5.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    # OCR
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "pillow>=10.2.0",
    # Audio
    "openai-whisper>=20231117",
    "pydub>=0.25.1",
    # NLP
    "spacy>=3.7.0",
    "nltk>=3.8.0",
    # Utilities
    "python-jose[cryptography]>=3.3.0",
    "redis>=5.0.0",
    "boto3>=1.34.0",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "Q"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
